chore: move schemas to oas folder
