+++ packages/salesforcedx-vscode-apex/src/oas/generationStrategy/wholeClassStrategy.ts
@@ -12,7 +12,7 @@ import {
   ApexClassOASGatherContextResponse,
   PromptGenerationResult,
   PromptGenerationStrategyBid
-} from '../../openApiUtilities/schemas';
+} from '../schemas';
 import { IMPOSED_FACTOR, PROMPT_TOKEN_MAX_LIMIT, SUM_TOKEN_MAX_LIMIT } from '.';
 import { GenerationStrategy } from './generationStrategy';
 import { prompts } from './prompts';
