+++ packages/salesforcedx-vscode-apex/src/oas/promptGenerationOrchestrator.ts
@@ -9,7 +9,7 @@ import {
   ApexClassOASEligibleResponse,
   ApexClassOASGatherContextResponse,
   PromptGenerationStrategyBid
-} from '../openApiUtilities/schemas';
+} from './schemas';
 import {
   GenerationStrategyFactory,
   Strategy,
