+++ packages/salesforcedx-vscode-apex/test/jest/commands/metadataOrchestrator.test.ts
@@ -10,7 +10,7 @@ import { ApexLanguageClient } from '../../../src/apexLanguageClient';
 import { MetadataOrchestrator } from '../../../src/commands/metadataOrchestrator';
 import { languageClientUtils } from '../../../src/languageUtils';
 import { nls } from '../../../src/messages';
-import { ApexOASResource } from '../../../src/openApiUtilities/schemas';
+import { ApexOASResource } from '../../../src/oas/schemas';
 import { getTelemetryService } from '../../../src/telemetry/telemetry';
 import { MockTelemetryService } from '../telemetry/mockTelemetryService';
